requirements.txt
pytest==7.1.2
allure-pytest==2.9.45
webdriver-manager==3.5.4
selenium==4.1.5
